pyproject.toml
[project]
name = "llmtitle"
version = "0.1.0"
description = "AI-Powered File Renamer using Google Gemini."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.5.0",
    "pypdf>=4.0.0",
    "tinytag>=2.1.2",
]

[project.scripts]
llmtitle = "llmtitle.cli:main"

[project.optional-dependencies]
test = ["pytest", "unittest-xml-reporting"]

[tool.setuptools.packages.find]
where = ["src"]
